@@ -1,55 +1,33 @@
 {
-  "_args": [
-    [
-      "react-dom@17.0.1",
-      "C:\\Users\\Chris\\Documents\\GitHub\\ValetProUser\\Client"
-    ]
+  "name": "react-dom",
+  "version": "18.2.0",
+  "description": "React package for working with the DOM.",
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/facebook/react.git",
+    "directory": "packages/react-dom"
+  },
+  "keywords": [
+    "react"
   ],
-  "_from": "react-dom@17.0.1",
-  "_id": "react-dom@17.0.1",
-  "_inBundle": false,
-  "_integrity": "sha512-6eV150oJZ9U2t9svnsspTMrWNyHc6chX0KzDeAOXftRa8bNeOKTTfCJ7KorIwenkHd2xqVTBTCZd79yk/lx/Ug==",
-  "_location": "/react-dom",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "react-dom@17.0.1",
-    "name": "react-dom",
-    "escapedName": "react-dom",
-    "rawSpec": "17.0.1",
-    "saveSpec": null,
-    "fetchSpec": "17.0.1"
-  },
-  "_requiredBy": [
-    "/"
-  ],
-  "_resolved": "https://registry.npmjs.org/react-dom/-/react-dom-17.0.1.tgz",
-  "_spec": "17.0.1",
-  "_where": "C:\\Users\\Chris\\Documents\\GitHub\\ValetProUser\\Client",
-  "browser": {
-    "./server.js": "./server.browser.js",
-    "./unstable-fizz.js": "./unstable-fizz.browser.js"
-  },
-  "browserify": {
-    "transform": [
-      "loose-envify"
-    ]
-  },
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/facebook/react/issues"
   },
+  "homepage": "https://reactjs.org/",
   "dependencies": {
     "loose-envify": "^1.1.0",
-    "object-assign": "^4.1.1",
-    "scheduler": "^0.20.1"
+    "scheduler": "^0.23.0"
   },
-  "description": "React package for working with the DOM.",
+  "peerDependencies": {
+    "react": "^18.2.0"
+  },
   "files": [
     "LICENSE",
     "README.md",
-    "build-info.json",
     "index.js",
+    "client.js",
     "profiling.js",
     "server.js",
     "server.browser.js",
@@ -58,23 +36,27 @@
     "cjs/",
     "umd/"
   ],
-  "homepage": "https://reactjs.org/",
-  "keywords": [
-    "react"
-  ],
-  "license": "MIT",
-  "main": "index.js",
-  "name": "react-dom",
-  "peerDependencies": {
-    "react": "17.0.1"
+  "exports": {
+    ".": "./index.js",
+    "./client": "./client.js",
+    "./server": {
+      "deno": "./server.browser.js",
+      "worker": "./server.browser.js",
+      "browser": "./server.browser.js",
+      "default": "./server.node.js"
+    },
+    "./server.browser": "./server.browser.js",
+    "./server.node": "./server.node.js",
+    "./profiling": "./profiling.js",
+    "./test-utils": "./test-utils.js",
+    "./package.json": "./package.json"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/facebook/react.git",
-    "directory": "packages/react-dom"
+  "browser": {
+    "./server.js": "./server.browser.js"
   },
-  "scripts": {
-    "start": "node server.js"
-  },
-  "version": "17.0.1"
+  "browserify": {
+    "transform": [
+      "loose-envify"
+    ]
+  }
 }
