@@ -10,10 +10,9 @@
     "bootstrap": "^4.6.1",
     "dayjs": "^1.10.4",
     "lodash": "^4.17.21",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-date-picker": "^9.0.0",
-    "react-dom": "^17.0.1",
-    "react-hook-form": "^6.14.2",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "4.0.1",
