@@ -1,5 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { createRoot } from "react-dom/client";
+
 import App from './App';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware, compose } from 'redux';
@@ -8,14 +10,16 @@ import reducer from './_reducers/index';
 import {BrowserRouter } from "react-router-dom";
 
 const store = createStore(reducer, compose(applyMiddleware(Thunk)))
-ReactDOM.render(
+
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
  
     <Provider store={store}>
       <BrowserRouter>
       <App />
       </BrowserRouter>
     </Provider>
-  ,
-  document.getElementById('root')
+  
 );
 
